backend/requirements.txt
django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
google-generativeai>=0.8.3
python-dotenv>=1.0
pillow>=10.2
dj-database-url>=2.2
psycopg[binary]>=3.2
gunicorn>=22.0
whitenoise>=6.7
